eval on held-out test data

+++ viz.py
@@ -38,7 +38,7 @@ def plot_confusion_matrix(y_true, y_pred, classes,
     # Compute confusion matrix
     cm = metrics.confusion_matrix(y_true, y_pred)
     # Only use the labels that appear in the data
-    classes = classes[unique_labels(y_true, y_pred)]
+    classes = classes[unique_labels(y_true.astype(np.int), y_pred.astype(np.int))]
     if normalize:
         cm = cm.astype('float') / cm.sum(axis=1)[:, np.newaxis]
